move match appointments to widgets
